buyer profile edit

+++ database/migrations/2019_04_01_135850_create_seller_profile_table.php
@@ -40,7 +40,7 @@ class CreateSellerProfileTable extends Migration
             $table->text('images')->nullable();;
             $table->integer('images_cnt')->nullable();;
 
-            $table->integer('status')->default(0);
+            $table->smallInteger('status')->default(0);
 
             $table->timestamps();
         });
